Handle hyphenated words properly.
Because the regex stopped consuming when it hit a single hyphen, we
couldn't handle words with hyphens. The intention was to stop consuming
only when we hit two hyphens in a row. But I don't know how to do that
with regex, or even if you can. The thing with comments though is that
they are terminated by the end of the line. So if you have a line that
starts as text and then has a comment, you know there's no more text on
that line. So we can simply consume all the comment data and then chop
it off after, which is what this does now.

We still need the branch that checks for double hyphens for comments
that are on their own line.