Ignore dist and build only once when server starts
We don't need to build before every request because changing anything
will restart the server, which will build.

When we didn't ignore the dist folder, the server kept restarting
unnecessarily because dist was changing on every request (or after the
other change, on startup). This should be more stable now.