Move folder removal step
Whenever ruby code changed we trashed the folder, which meant that we
deleted our compiled JS. Because no typescript changed, wepback didn't
recompile it.

Folder removal is the start of the whole process, not specific to the
ruby build step, so it shouldn't be part of the ruby build step.