Use fractional gem instead of custom method
fractional gem might have ugly code but I don't have to see it! and yes
this is really only for if JS is disabled, so no one will ever see it.
But it's barely any overhead. As this is a static site the only possible
performance hit could be in the build step, and it's still very fast.